requirements.txt
torch>=2.2
torchvision>=0.17
transformers>=4.40
accelerate>=0.28
scikit-learn>=1.4
pandas>=2.2
numpy>=1.26
Pillow>=10.0
tqdm>=4.66
matplotlib>=3.8
pytest>=8.0
